bugfix: self-modifications to versions were not in right directories
There is no "PUBLICVERSION_*_NEW" (that was a bug fixed long ago). So the proposed version.cmake, README.md and CHANGES.md were meant to go in the proposed "VERSION_*_NEW" directories.

Ditto the proposed versions (VERSION_*_NEW, not PUBLICVERSION_*_NEW) should be embedded in the version.cmake, README.md and CHANGES.md