bugfix: Parse subtree split rather than use branch
The `git subtree split --rejoin --branch <x>` fails if <x> exists, so deleting the <x> branch if it existed was the previous behavior. But there was no creation <x> if there was no change to the rejoin commit. So no way to use --rejoin --branch consistently.

Instead just use --rejoin and parse the commit id; then tag that.